+++ src/main/java/io/github/astrapi69/xstream/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package provides classes for the transformation of objects to xml and back with xstream
+ *
+ * @author Asterios Raptis
+ */
+package io.github.astrapi69.xstream;