+++ src/test/java/io/github/astrapi69/xstream/package-info.java
@@ -1,0 +1,1 @@
+package io.github.astrapi69.xstream;