+++ CHANGELOG.md
@@ -4,14 +4,20 @@
 Version 1.1-SNAPSHOT
 -------------
 
+ADDED:
+
+- all required package-info.java files
+- new converter class that transforms a xml file to a java object
+- new extension class that transforms a xml file to a java object
+
 CHANGED:
 
-- update gradle to new version 7.6-rc-1
+- update gradle to new version 7.6-rc-2
 - update of gradle plugin dependency com.github.ben-manes.versions.gradle.plugin to new minor version 0.43.0
 - update of gradle plugin dependency com.diffplug.spotless:spotless-plugin-gradle to new minor version 6.11.0
 - update of gradle plugin dependency io.freefair.gradle:lombok-plugin to new minor version 6.5.1
-- update of dependency xml-api to new minor version 1.3
-- update of dependency file-worker to new minor version 11.5
+- update of dependency xml-api to new minor version 1.4
+- update of dependency file-worker to new minor version 11.5.1
 - update of dependency jettison to new minor version 1.5.1
 - update of dependency throwable to new minor version 2.3
 - update of test dependency test-object to new minor version 7.2